+++ ProdManager/routes/scope/views.py
@@ -10,7 +10,7 @@ from ProdManager.helpers.form import strip_input
 from ProdManager.helpers.links import custom_url_for
 
 from ProdManager.models import (
-  Scope, Monitor, Incident, Maintenance,
+  Scope, Monitor, Incident, Maintenance, Announcement
 )
 
 from ProdManager.plugins import lang
@@ -111,6 +111,15 @@ def show(resource_id):
     ),
     past_maintenances_filters=Maintenance.past_filter(raw=True),
     monitors_count=scope.monitors_count(),
+    ongoing_announcements_filters=Announcement.ongoing_filter(raw=True),
+    ongoing_announcements=list_resources_from_query(
+      Announcement,
+      query=scope.announcements,
+      filters=Announcement.ongoing_filter(),
+      orders=Announcement.reverse_order(),
+      paginate=False,
+      limit=10,
+    ),
   ), 200